+++ OnlineAdmission.BLL/IManager/IStudentManager.cs
@@ -10,7 +10,7 @@ namespace OnlineAdmission.BLL.IManager
     public interface IStudentManager : IManager<Student>
     {
 
-        Task<Student> GetStudentByHSCRollAsync(int hscRoll);
+        Task<Student> GetStudentByHSCRollAsync(long hscRoll);
         Task<Student> GetStudentBySSCRollAsync(int hscRoll, string boardName);
         Task<int> GetCountAsync(int subjectCode);
         Task<Student> GetByAdmissionRollAsync(int NURoll);
